Add a nice way to create a builder for subqueries (#7040)
|      Q       |   A
|------------- | -----------
| Type         | feature
| Fixed issues | N/A

#### Summary

In DBAL 3, we could ask q query builder for its database connection.
That accessor has been removed in DBAL 4 which is a good thing.

Currently, I'm migrating a codebase that used the
`QueryBuilder::getConnection()` method in order to create a fresh
builder for constructing a subquery.

```php
$subQueryBuilder = $queryBuilder->getConnection()->createQueryBuilder();
```

I believe this is a valid requirement, especially given that we allow
query builders for unions and CTEs.

This PR proposes a new method `sub()` which creates a fresh builder for
the same connection, pretty much like the already existing method
`expr()` does for the expression builder.

```php
$subQueryBuilder = $queryBuilder->sub();
```